+++ CHANGELOG.md
@@ -3,7 +3,7 @@ CHANGELOG
 
 ## HEAD (Unreleased)
 - Fix refresh loop with self hosted backend [#194](https://github.com/pulumi/pulumi-kubernetes-operator/pull/194)
-
+- Improve workdir cleanup logic [#195](https://github.com/pulumi/pulumi-kubernetes-operator/pull/195)
 
 ## 0.0.19 (2021-09-02)
 - Add support for safe upgrades and graceful shutdowns [#189](https://github.com/pulumi/pulumi-kubernetes-operator/pull/189)